requirements.txt
roboflow
ultralytics
fastapi
uvicorn
dvc
apache-airflow
pillow
numpy
opencv-python
pyyaml